refactor: 유효성 검사 어노테이션 추가 및 MemberDTO 수정
